+++ rust/guest_wrapper/artifacts/chain_guest/CHANGELOG.md
@@ -4,3 +4,4 @@
   * `2dfd3adf61327dc1ad47a637f5771496458ad5dc9a10de41b05cb348fa5d2eee` – Use old guest IDs in chain guest [#1584](https://github.com/vlayer-xyz/vlayer/pull/1584)
   * `acbedbbeb664cdc1c58abd3d1c6c5a6e66d7cd206d752e3ddb7bf133f5f04ce7` – Use old guest IDs in chain host [#1603](https://github.com/vlayer-xyz/vlayer/pull/1603)
   * `d35081d72b7d4875780f69da408127523669dfb2f17a05c23dc2506b74a92c83` – Revert Risc0 v1.2.1 to v1.2.0
+  * `9c027beee25f4f98a7cf995b5d752222673c6743e1cc6f38839b3bc6445f3441` – Initial teleport implementation